Logging for Assertion error better in the JMX tests
